Made the drive train work- Yuhyun
i commented 90% of the code and now the drivetrain works :D
