--- infinite recharge refined/.wpilib/wpilib_preferences.json
@@ -1,6 +1,0 @@
-{
-    "enableCppIntellisense": false,
-    "currentLanguage": "java",
-    "projectYear": "2020",
-    "teamNumber": 2637
-}
