--- infinite recharge refined/vendordeps/REVRobotics.json
@@ -1,70 +1,0 @@
-{
-    "fileName": "REVRobotics.json",
-    "name": "REVRobotics",
-    "version": "1.5.1",
-    "uuid": "3f48eb8c-50fe-43a6-9cb7-44c86353c4cb",
-    "mavenUrls": [
-        "http://www.revrobotics.com/content/sw/max/sdk/maven/"
-    ],
-    "jsonUrl": "http://www.revrobotics.com/content/sw/max/sdk/REVRobotics.json",
-    "javaDependencies": [
-        {
-            "groupId": "com.revrobotics.frc",
-            "artifactId": "SparkMax-java",
-            "version": "1.5.1"
-        }
-    ],
-    "jniDependencies": [
-        {
-            "groupId": "com.revrobotics.frc",
-            "artifactId": "SparkMax-driver",
-            "version": "1.5.1",
-            "skipInvalidPlatforms": true,
-            "isJar": false,
-            "validPlatforms": [
-                "windowsx86-64",
-                "windowsx86",
-                "linuxaarch64bionic",
-                "linuxx86-64",
-                "linuxathena",
-                "linuxraspbian"
-            ]
-        }
-    ],
-    "cppDependencies": [
-        {
-            "groupId": "com.revrobotics.frc",
-            "artifactId": "SparkMax-cpp",
-            "version": "1.5.1",
-            "libName": "SparkMax",
-            "headerClassifier": "headers",
-            "sharedLibrary": false,
-            "skipInvalidPlatforms": true,
-            "binaryPlatforms": [
-                "windowsx86-64",
-                "windowsx86",
-                "linuxaarch64bionic",
-                "linuxx86-64",
-                "linuxathena",
-                "linuxraspbian"
-            ]
-        },
-        {
-            "groupId": "com.revrobotics.frc",
-            "artifactId": "SparkMax-driver",
-            "version": "1.5.1",
-            "libName": "SparkMaxDriver",
-            "headerClassifier": "headers",
-            "sharedLibrary": false,
-            "skipInvalidPlatforms": true,
-            "binaryPlatforms": [
-                "windowsx86-64",
-                "windowsx86",
-                "linuxaarch64bionic",
-                "linuxx86-64",
-                "linuxathena",
-                "linuxraspbian"
-            ]
-        }
-    ]
-}
