--- infinite recharge refined/vendordeps/WPILibOldCommands.json
@@ -1,37 +1,0 @@
-{
-  "fileName": "WPILibOldCommands.json",
-  "name": "WPILib-Old-Commands",
-  "version": "2020.0.0",
-  "uuid": "b066afc2-5c18-43c4-b758-43381fcb275e",
-  "mavenUrls": [],
-  "jsonUrl": "",
-  "javaDependencies": [
-      {
-          "groupId": "edu.wpi.first.wpilibOldCommands",
-          "artifactId": "wpilibOldCommands-java",
-          "version": "wpilib"
-      }
-  ],
-  "jniDependencies": [],
-  "cppDependencies": [
-      {
-          "groupId": "edu.wpi.first.wpilibOldCommands",
-          "artifactId": "wpilibOldCommands-cpp",
-          "version": "wpilib",
-          "libName": "wpilibOldCommands",
-          "headerClassifier": "headers",
-          "sourcesClassifier": "sources",
-          "sharedLibrary": true,
-          "skipInvalidPlatforms": true,
-          "binaryPlatforms": [
-              "linuxathena",
-              "linuxraspbian",
-              "linuxaarch64bionic",
-              "windowsx86-64",
-              "windowsx86",
-              "linuxx86-64",
-              "osxx86-64"
-          ]
-      }
-  ]
-}
